swc_ecma_parser::input

Trait Tokens

pub trait Tokens: Clone {
    type Checkpoint;

    // Required methods
    fn set_ctx(&mut self, ctx: Context);
    fn ctx(&self) -> Context;
    fn ctx_mut(&mut self) -> &mut Context;
    fn syntax(&self) -> SyntaxFlags;
    fn target(&self) -> EsVersion;
    fn checkpoint_save(&self) -> Self::Checkpoint;
    fn checkpoint_load(&mut self, checkpoint: Self::Checkpoint);
    fn read_string(&self, span: Span) -> &str;
    fn set_expr_allowed(&mut self, allow: bool);
    fn set_next_regexp(&mut self, start: Option<BytePos>);
    fn add_error(&mut self, error: Error);
    fn add_module_mode_error(&mut self, error: Error);
    fn end_pos(&self) -> BytePos;
    fn take_errors(&mut self) -> Vec<Error>;
    fn take_script_module_errors(&mut self) -> Vec<Error>;
    fn update_token_flags(&mut self, f: impl FnOnce(&mut TokenFlags));
    fn token_flags(&self) -> TokenFlags;
    fn clone_token_value(&self) -> Option<TokenValue>;
    fn take_token_value(&mut self) -> Option<TokenValue>;
    fn get_token_value(&self) -> Option<&TokenValue>;
    fn set_token_value(&mut self, token_value: Option<TokenValue>);
    fn first_token(&mut self) -> TokenAndSpan;
    fn next_token(&mut self) -> TokenAndSpan;
    fn scan_jsx_token(&mut self) -> TokenAndSpan;
    fn scan_jsx_open_el_terminal_token(&mut self) -> TokenAndSpan;
    fn rescan_jsx_open_el_terminal_token(
        &mut self,
        reset: BytePos,
    ) -> TokenAndSpan;
    fn rescan_jsx_token(&mut self, reset: BytePos) -> TokenAndSpan;
    fn scan_jsx_identifier(&mut self, start: BytePos) -> TokenAndSpan;
    fn scan_jsx_attribute_value(&mut self) -> TokenAndSpan;
    fn rescan_template_token(
        &mut self,
        start: BytePos,
        start_with_back_tick: bool,
    ) -> TokenAndSpan;

    // Provided methods
    fn start_pos(&self) -> BytePos { ... }
    fn has_errors(&self) -> bool { ... }
    fn diagnostic_checkpoint_save(&self) -> (usize, usize) { ... }
    fn diagnostic_checkpoint_load(&mut self, _checkpoint: (usize, usize)) { ... }
    fn set_defer_comments(&mut self, _defer: bool) { ... }
    fn finalize_comments(&mut self) { ... }

}

Clone should be cheap if you are parsing typescript because typescript syntax requires backtracking.

Required Associated Types

type Checkpoint

Required Methods

fn set_ctx(&mut self, ctx: Context)

fn ctx(&self) -> Context

fn ctx_mut(&mut self) -> &mut Context

fn syntax(&self) -> SyntaxFlags

fn target(&self) -> EsVersion

fn checkpoint_save(&self) -> Self::Checkpoint

fn checkpoint_load(&mut self, checkpoint: Self::Checkpoint)

fn read_string(&self, span: Span) -> &str

fn set_expr_allowed(&mut self, allow: bool)

fn set_next_regexp(&mut self, start: Option<BytePos>)

fn add_error(&mut self, error: Error)

Implementors should use Rc<RefCell<Vec>>.

It is required because parser should backtrack while parsing typescript code.

fn add_module_mode_error(&mut self, error: Error)

Add an error which is valid syntax in script mode.

This errors should be dropped if it’s not a module.

Implementor should check for if Context.module, and buffer errors if module is false. Also, implementors should move errors to the error buffer on set_ctx if the parser mode become module mode.

fn end_pos(&self) -> BytePos

fn take_errors(&mut self) -> Vec<Error>

fn take_script_module_errors(&mut self) -> Vec<Error>

If the program was parsed as a script, this contains the module errors should the program be identified as a module in the future.

fn update_token_flags(&mut self, f: impl FnOnce(&mut TokenFlags))

fn token_flags(&self) -> TokenFlags

fn clone_token_value(&self) -> Option<TokenValue>

fn take_token_value(&mut self) -> Option<TokenValue>

fn get_token_value(&self) -> Option<&TokenValue>

fn set_token_value(&mut self, token_value: Option<TokenValue>)

fn first_token(&mut self) -> TokenAndSpan

Returns the first token in the file.

This function should only be called at the first time of bump. It’s mainly used for shebang.

fn next_token(&mut self) -> TokenAndSpan

Returns the next token from the input stream.

This method always returns a TokenAndSpan. When the end of input is reached, it returns Token::Eof, and subsequent calls will continue returning Token::Eof.

fn scan_jsx_token(&mut self) -> TokenAndSpan

fn scan_jsx_open_el_terminal_token(&mut self) -> TokenAndSpan

fn rescan_jsx_open_el_terminal_token(&mut self, reset: BytePos) -> TokenAndSpan

fn rescan_jsx_token(&mut self, reset: BytePos) -> TokenAndSpan

fn scan_jsx_identifier(&mut self, start: BytePos) -> TokenAndSpan

fn scan_jsx_attribute_value(&mut self) -> TokenAndSpan

fn rescan_template_token( &mut self, start: BytePos, start_with_back_tick: bool, ) -> TokenAndSpan

Provided Methods

fn start_pos(&self) -> BytePos

fn has_errors(&self) -> bool

Returns whether parsing has emitted a non-module-mode error.

Custom token sources which buffer recoverable errors should override this together with the diagnostic checkpoint methods.

fn diagnostic_checkpoint_save(&self) -> (usize, usize)

Saves the diagnostic buffer lengths for a whole-program retry.

fn diagnostic_checkpoint_load(&mut self, _checkpoint: (usize, usize))

Discards diagnostics emitted after a whole-program checkpoint.

fn set_defer_comments(&mut self, _defer: bool)

Defers publishing buffered comments while a parse may be rewound.

fn finalize_comments(&mut self)

Publishes all buffered comments after the parser commits a parse.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".