swc_ecma_parser
Module lexer

ECMAScript lexer.

Modules

input

util

Ported from babylon/util/identifier.js

Structs

Lexer

TokenContexts

Enums

TokenContext

The algorithm used to determine whether a regexp can appear at a given point in the program is loosely based on sweet.js’ approach. See https://github.com/mozilla/sweet.js/wiki/design

Traits

Input