Struct swc_ecma_parser::lexer::input::StringInput

pub struct StringInput<'a> { /* fields omitted */ }

Implementation of Input.

Implementations

impl<'a> StringInput<'a>

pub fn new(src: &'a str, start: BytePos, end: BytePos) -> StringInput<'a>

start and end can be arbitrary value, but start should be less than or equal to end.

swc get this value from [SourceMap] because code generator depends on some methods of [SourceMap]. If you are not going to use methods from [SourceMap], you may use any value.