Trait swc_ecma_parser::lexer::Input

pub trait Input: Clone {

    fn cur(&mut self) -> Option<char>;
    fn peek(&mut self) -> Option<char>;
    fn peek_ahead(&mut self) -> Option<char>;
    fn bump(&mut self);
    fn is_at_start(&self) -> bool;
    fn cur_pos(&mut self) -> BytePos;
    fn last_pos(&self) -> BytePos;
    fn slice(&mut self, start: BytePos, end: BytePos) -> &str;
    fn uncons_while<F>(&mut self, f: F) -> &str
    where
        F: FnMut(char) -> bool;
    fn find<F>(&mut self, f: F) -> Option<BytePos>
    where
        F: FnMut(char) -> bool;
    fn reset_to(&mut self, to: BytePos);
    fn is_str(&self, s: &str) -> bool;

    fn cur_as_ascii(&mut self) -> Option<u8> { ... }
    fn is_byte(&mut self, c: u8) -> bool { ... }
    fn eat_byte(&mut self, c: u8) -> bool { ... }

}

Required Methods

fn cur(&mut self) -> Option<char>

fn peek(&mut self) -> Option<char>

fn peek_ahead(&mut self) -> Option<char>

fn bump(&mut self)

fn is_at_start(&self) -> bool

fn cur_pos(&mut self) -> BytePos

fn last_pos(&self) -> BytePos

fn slice(&mut self, start: BytePos, end: BytePos) -> &str

fn uncons_while<F>(&mut self, f: F) -> &strwhere
F: FnMut(char) -> bool,

Takes items from stream, testing each one with predicate. returns the range of items which passed predicate.

fn find<F>(&mut self, f: F) -> Option<BytePos>where
F: FnMut(char) -> bool,

This method modifies [last_pos()] and [cur_pos()].

fn reset_to(&mut self, to: BytePos)

fn is_str(&self, s: &str) -> bool

Implementors can override the method to make it faster.

s must be ASCII only.

Provided Methods

fn cur_as_ascii(&mut self) -> Option<u8>

Returns None if it’s end of input or current character is not an ascii character.

fn is_byte(&mut self, c: u8) -> bool

Implementors can override the method to make it faster.

c must be ASCII.

fn eat_byte(&mut self, c: u8) -> bool

Implementors can override the method to make it faster.

c must be ASCII.