Trait swc_ecma_parser::lexer::util::CharExt

pub trait CharExt: Copy {
    fn to_char(self) -> Option<char>;

    fn is_ident_start(self) -> bool { ... }

    fn is_ident_part(self) -> bool { ... }

    fn is_line_break(self) -> bool { ... }

    fn is_ws(self) -> bool { ... }
}

Implemented for char.

Required Methods

fn to_char(self) -> Option<char>

Provided Methods

fn is_ident_start(self) -> bool

Test whether a given character code starts an identifier.

https://tc39.github.io/ecma262/#prod-IdentifierStart

fn is_ident_part(self) -> bool

Test whether a given character is part of an identifier.

fn is_line_break(self) -> bool

See https://tc39.github.io/ecma262/#sec-line-terminators

fn is_ws(self) -> bool

See https://tc39.github.io/ecma262/#sec-white-space