Struct swc_ecma_parser::parser::Parser

pub struct Parser<'a, I: Input> {
    session: Session<'a>,
    state: State,
    input: ParserInput<'a, I>,
}

EcmaScript parser.

Fields
session: Session<'a> state: State input: ParserInput<'a, I>
Methods

impl<'a, I: Input> Parser<'a, I>

fn parse_async_fn_expr(&mut self) -> PResult<'a, Box<Expr>>

fn parse_fn_expr(&mut self) -> PResult<'a, Box<Expr>>

Parse function expression

fn parse_async_fn_decl(&mut self) -> PResult<'a, Decl>

fn parse_fn_decl(&mut self) -> PResult<'a, Decl>

fn parse_default_async_fn(&mut self) -> PResult<'a, ExportDefaultDecl>

fn parse_default_fn(&mut self) -> PResult<'a, ExportDefaultDecl>

fn parse_class_decl(&mut self) -> PResult<'a, Decl>

fn parse_class_expr(&mut self) -> PResult<'a, Box<Expr>>

fn parse_default_class(&mut self) -> PResult<'a, ExportDefaultDecl>

fn parse_class<T>(&mut self) -> PResult<'a, T> where
T: OutputType,
Self: MaybeOptionalIdentParser<'a, T::Ident>,

fn parse_class_body(&mut self) -> PResult<'a, Vec<ClassMethod>>

fn parse_class_element(&mut self) -> PResult<'a, ClassMethod>

fn parse_fn<T>(&mut self, start_of_async: Option<BytePos>) -> PResult<'a, T> where
T: OutputType,
Self: MaybeOptionalIdentParser<'a, T::Ident>,

fn parse_fn_args_body<F>(
&mut self,
start: BytePos,
parse_args: F,
is_async: bool,
is_generator: bool
) -> PResult<'a, Function> where
F: FnOnce(&mut Self) -> PResult<'a, Vec<Pat>>,

parse_args closure should not eat '(' or ')'.

fn parse_method_def(
&mut self,
start_of_static: Option<BytePos>
) -> PResult<'a, ClassMethod>

fn parse_fn_body<T>(
&mut self,
is_async: bool,
is_generator: bool
) -> PResult<'a, T> where
Self: FnBodyParser<'a, T>,

impl<'a, I: Input> Parser<'a, I>

fn parse_object<T>(&mut self) -> PResult<'a, T> where
Self: ParseObject<'a, T>,

Parse a object literal or object pattern.

fn parse_prop_name(&mut self) -> PResult<'a, PropName>

spec: 'PropertyName'

impl<'a, I: Input> Parser<'a, I>

fn parse_bin_expr(&mut self) -> PResult<'a, Box<Expr>>

Name from spec: 'LogicalORExpression'

fn parse_bin_op_recursively(
&mut self,
left: Box<Expr>,
min_prec: u8
) -> PResult<'a, Box<Expr>>

Parse binary operators with the operator precedence parsing algorithm. left is the left-hand side of the operator. minPrec provides context that allows the function to stop and defer further parser to one of its callers when it encounters an operator that has a lower precedence than the set it is parsing.

fn parse_unary_expr(&mut self) -> PResult<'a, Box<Expr>>

Parse unary expression and update expression.

spec: 'UnaryExpression'

fn parse_await_expr(&mut self) -> PResult<'a, Box<Expr>>

impl<'a, I: Input> Parser<'a, I>

fn verify_expr(&self, expr: Box<Expr>) -> PResult<'a, Box<Expr>>

impl<'a, I: Input> Parser<'a, I>

pub fn parse_expr(&mut self) -> PResult<'a, Box<Expr>>

fn parse_assignment_expr(&mut self) -> PResult<'a, Box<Expr>>

Parse an assignment expression. This includes applications of operators like +=.

fn parse_cond_expr(&mut self) -> PResult<'a, Box<Expr>>

Spec: 'ConditionalExpression'

fn parse_primary_expr(&mut self) -> PResult<'a, Box<Expr>>

Parse a primary expression or arrow function

fn parse_array_lit(&mut self) -> PResult<'a, Box<Expr>>

fn parse_member_expr(&mut self) -> PResult<'a, Box<Expr>>

fn parse_member_expr_or_new_expr(
&mut self,
is_new_expr: bool
) -> PResult<'a, Box<Expr>>

is_new_expr: true iff we are parsing production 'NewExpression'.

fn parse_new_expr(&mut self) -> PResult<'a, Box<Expr>>

Parse NewExpresion. This includes MemberExpression.

fn parse_args(&mut self) -> PResult<'a, Vec<ExprOrSpread>>

Parse Arguments[Yield, Await]

fn parse_expr_or_spread(&mut self) -> PResult<'a, ExprOrSpread>

AssignmentExpression[+In, ?Yield, ?Await] ...AssignmentExpression[+In, ?Yield, ?Await]

fn parse_paren_expr_or_arrow_fn(
&mut self,
can_be_arrow: bool
) -> PResult<'a, Box<Expr>>

Parse paren expression or arrow function expression.

fn parse_tpl_lit(&mut self, tag: Option<Box<Expr>>) -> PResult<'a, TplLit>

fn parse_tpl_element(&mut self, is_tagged: bool) -> PResult<'a, TplElement>

fn parse_subscripts(
&mut self,
obj: ExprOrSuper,
no_call: bool
) -> PResult<'a, Box<Expr>>

fn parse_subscript(
&mut self,
obj: ExprOrSuper,
no_call: bool
) -> PResult<'a, (Box<Expr>, bool)>

returned bool is true if this method should be called again.

fn parse_lhs_expr(&mut self) -> PResult<'a, Box<Expr>>

Parse call, dot, and []-subscript expressions.

fn parse_expr_or_pat(&mut self) -> PResult<'a, Box<Expr>>

fn parse_args_or_pats(&mut self) -> PResult<'a, Vec<ExprOrSpread>>

impl<'a, I: Input> Parser<'a, I>

simple leaf methods.

fn parse_yield_expr(&mut self) -> PResult<'a, Box<Expr>>

fn parse_lit(&mut self) -> PResult<'a, Lit>

12.2.5 Array Initializer

impl<'a, I: Input> Parser<'a, I>

fn parse_ident_ref(&mut self) -> PResult<'a, Ident>

IdentifierReference

fn parse_label_ident(&mut self) -> PResult<'a, Ident>

LabelIdentifier

fn parse_ident_name(&mut self) -> PResult<'a, Ident>

Use this when spec says "IdentiferName". This allows idents like catch.

fn parse_ident(
&mut self,
incl_yield: bool,
incl_await: bool
) -> PResult<'a, Ident>

Identifier

In strict mode, "yield" is SyntaxError if matched.

impl<'a, I: Input> Parser<'a, I>

fn parse_import(&mut self) -> PResult<'a, ModuleDecl>

fn parse_import_specifier(&mut self) -> PResult<'a, ImportSpecifier>

Parse foo, foo2 as bar in import { foo, foo2 as bar }

fn parse_imported_default_binding(&mut self) -> PResult<'a, Ident>

fn parse_imported_binding(&mut self) -> PResult<'a, Ident>

fn parse_export(&mut self) -> PResult<'a, ModuleDecl>

fn parse_export_specifier(&mut self) -> PResult<'a, ExportSpecifier>

fn parse_from_clause_and_semi(&mut self) -> PResult<'a, String>

impl<'a, I: Input> Parser<'a, I>

fn parse_block_body<Type>(
&mut self,
allow_directives: bool,
top_level: bool,
end: Option<&Token>
) -> PResult<'a, Vec<Type>> where
Self: StmtLikeParser<'a, Type>,
Type: IsDirective + From<Stmt>,

fn parse_stmt(&mut self, top_level: bool) -> PResult<'a, Stmt>

fn parse_stmt_list_item(&mut self, top_level: bool) -> PResult<'a, Stmt>

fn parse_stmt_like<Type>(
&mut self,
include_decl: bool,
top_level: bool
) -> PResult<'a, Type> where
Self: StmtLikeParser<'a, Type>,
Type: IsDirective + From<Stmt>,

Parse a statement, declaration or module item.

fn parse_stmt_internal(
&mut self,
include_decl: bool,
top_level: bool
) -> PResult<'a, Stmt>

fn parse_if_stmt(&mut self) -> PResult<'a, Stmt>

fn parse_return_stmt(&mut self) -> PResult<'a, Stmt>

fn parse_switch_stmt(&mut self) -> PResult<'a, Stmt>

fn parse_throw_stmt(&mut self) -> PResult<'a, Stmt>

fn parse_try_stmt(&mut self) -> PResult<'a, Stmt>

fn parse_catch_param(&mut self) -> PResult<'a, Pat>

fn parse_var_stmt(&mut self, for_loop: bool) -> PResult<'a, VarDecl>

fn parse_var_declarator(&mut self, for_loop: bool) -> PResult<'a, VarDeclarator>

fn parse_do_stmt(&mut self) -> PResult<'a, Stmt>

fn parse_while_stmt(&mut self) -> PResult<'a, Stmt>

fn parse_with_stmt(&mut self) -> PResult<'a, Stmt>

fn parse_block(&mut self, allow_directives: bool) -> PResult<'a, BlockStmt>

fn parse_labelled_stmt(&mut self, label: Ident) -> PResult<'a, Stmt>

fn parse_for_stmt(&mut self) -> PResult<'a, Stmt>

fn parse_for_head(&mut self) -> PResult<'a, ForHead>

fn parse_for_each_head(&mut self, left: VarDeclOrPat) -> PResult<'a, ForHead>

fn parse_normal_for_head(
&mut self,
init: Option<VarDeclOrExpr>
) -> PResult<'a, ForHead>

impl<'a, I: Input> Parser<'a, I>

fn parse_opt_binding_ident(&mut self) -> PResult<'a, Option<Ident>>

fn parse_binding_ident(&mut self) -> PResult<'a, Ident>

babel: parseBindingIdentifier

spec: BindingIdentifier

fn parse_binding_pat_or_ident(&mut self) -> PResult<'a, Pat>

fn parse_binding_element(&mut self) -> PResult<'a, Pat>

babel: parseBindingAtom

fn parse_array_binding_pat(&mut self) -> PResult<'a, Pat>

fn parse_formal_param(&mut self) -> PResult<'a, Pat>

spec: 'FormalParameter'

fn parse_formal_params(&mut self) -> PResult<'a, Vec<Pat>>

spec: 'FormalParameterList'

fn parse_unique_formal_params(&mut self) -> PResult<'a, Vec<Pat>>

impl<'a, I: Input> Parser<'a, I>

fn reparse_expr_as_pat(
&mut self,
pat_ty: PatType,
expr: Box<Expr>
) -> PResult<'a, Pat>

This does not return 'rest' pattern because non-last parameter cannot be rest.

fn parse_exprs_as_params(
&mut self,
exprs: Vec<ExprOrSpread>
) -> PResult<'a, Vec<Pat>>

impl<'a, I: Input> Parser<'a, I>

fn with_ctx<'w>(&'w mut self, ctx: Context) -> WithCtx<'w, 'a, I>

Original context is restored when returned guard is dropped.

fn set_ctx(&mut self, ctx: Context)

fn strict_mode<'w>(&'w mut self) -> WithCtx<'w, 'a, I>

fn include_in_expr<'w>(
&'w mut self,
include_in_expr: bool
) -> WithCtx<'w, 'a, I>

Original context is restored when returned guard is dropped.

fn parse_with<F, Ret>(&mut self, f: F) -> Ret where
F: FnOnce(&mut Self) -> Ret,

Parse with given closure

fn spanned<F, Node, Ret>(&mut self, f: F) -> PResult<'a, Node> where
F: FnOnce(&mut Self) -> PResult<'a, Ret>,
Node: Spanned<Ret>,

impl<'a, I: Input> Parser<'a, I>

pub fn new(session: Session<'a>, input: I) -> Self

pub fn parse_script(&mut self) -> PResult<'a, Vec<Stmt>>

pub fn parse_module(&mut self) -> PResult<'a, Module>

const fn ctx(&self) -> Context