swc_ecma_parser::lexer::util
Trait CharExt

pub trait CharExt: Copy {
    // Required method
    fn to_char(self) -> Option<char>;

    // Provided methods
    fn is_ident_start(self) -> bool { ... }
    fn is_ident_part(self) -> bool { ... }
    fn is_line_terminator(self) -> bool { ... }
    fn is_line_break(self) -> bool { ... }
    fn is_ws(self) -> bool { ... }
}

Implemented for char.

Required Methods

fn to_char(self) -> Option<char>

Provided Methods

fn is_ident_start(self) -> bool

Test whether a given character code starts an identifier.

https://tc39.github.io/ecma262/#prod-IdentifierStart

fn is_ident_part(self) -> bool

Test whether a given character is part of an identifier.

fn is_line_terminator(self) -> bool

See https://tc39.github.io/ecma262/#sec-line-terminators

fn is_line_break(self) -> bool

See https://tc39.github.io/ecma262/#sec-literals-string-literals

fn is_ws(self) -> bool

See https://tc39.github.io/ecma262/#sec-white-space

Object Safety

This trait is not object safe.